react_dev: removed big part of useless code, moved all events on index, and structured
