+++ react/package.json
@@ -7,7 +7,6 @@
     "apollo-cache-inmemory": "^1.6.5",
     "apollo-client": "^2.6.8",
     "graphql": "^14.5.4",
-    "leadlionpartnerro.themarketer": "https://razvan--leadlionpartnerro.myvtex.com/_v/private/typings/linked/v1/leadlionpartnerro.themarketer@0.1.3+build1676640828/public/@types/leadlionpartnerro.themarketer",
     "promise-fn-retry": "^1.0.10",
     "react-intl": "^3.11.0",
     "typescript": "3.9.7",
